refactor(execution): use public API imports instead of submodule imports
Replace 57 submodule-level external imports with package-level public
API imports across execution module. Symbols without public API access
(11 unique symbols) remain as internal imports and will have dependency
issues filed separately.

Changes:
- Replace `from vibe3.clients.sqlite_client import SQLiteClient` with
  `from vibe3.clients import SQLiteClient`
- Replace `from vibe3.models.execution_request import ExecutionRequest`
  with `from vibe3.models import ExecutionRequest`
- Similar replacements for models, config, services, orchestra, agents,
  and environment imports
- Consolidate multiple imports from same package into single lines
- Fix import sorting with ruff format

Remaining internal imports (Category B, no public API available):
- SessionRegistryService (vibe3.environment.session_registry)
- VerdictRecord, VerdictValue (vibe3.models.verdict*)
- AgentOptions (vibe3.models.review_runner)
- get_role_output_contract, get_role_section (vibe3.config.role_policy)
- GOVERNANCE_GATE_CONFIG (vibe3.config.role_gates)
- sync_models_json (vibe3.agents.backends.codeagent_config)
- agent_preset symbols (vibe3.config.agent_preset)
- get_store (vibe3.clients.store_context)
- GitHubAPIError (vibe3.exceptions.runtime_errors)

All 140 execution module tests pass. No behavior changes.

Ref: #1922
