+++ src/vibe3/execution/codeagent_runner.py
@@ -9,8 +9,7 @@ from typing import TYPE_CHECKING, cast
 from loguru import logger
 from typer import echo
 
-from vibe3.clients.git_client import GitClient
-from vibe3.clients.sqlite_client import SQLiteClient
+from vibe3.clients import GitClient, SQLiteClient
 
 if TYPE_CHECKING:
     from loguru import Logger
@@ -21,8 +20,8 @@ if TYPE_CHECKING:
         CodeagentResult,
     )
     from vibe3.exceptions.error_severity import ErrorSeverity
+from vibe3.config import VibeConfig
 from vibe3.config.role_policy import get_role_section
-from vibe3.config.settings import VibeConfig
 from vibe3.execution.codeagent_support import resolve_command_agent_options
 from vibe3.execution.execution_lifecycle import (
     execution_prefix,
@@ -30,10 +29,9 @@ from vibe3.execution.execution_lifecycle import (
 )
 from vibe3.execution.noop_gate import apply_unified_noop_gate
 from vibe3.execution.session_service import load_session_id
-from vibe3.models.execution_request import ExecutionRequest
+from vibe3.models import ExecutionRequest
 from vibe3.models.review_runner import AgentOptions
-from vibe3.services.actor_support import format_agent_actor
-from vibe3.services.handoff_service import HandoffService
+from vibe3.services import HandoffService, format_agent_actor
 
 
 def _severity_event_type(role: str, severity: "ErrorSeverity") -> str:
@@ -96,8 +94,7 @@ class CodeagentExecutionService:
             log.info(f"Removed {handoff_label} label from #{issue_number}")
         except Exception as exc:
             log.warning(
-                f"Failed to remove {handoff_label} label from "
-                f"#{issue_number}: {exc}"
+                f"Failed to remove {handoff_label} label from #{issue_number}: {exc}"
             )
 
     @staticmethod
