+++ src/vibe3/execution/contracts.py
@@ -4,11 +4,11 @@ from pathlib import Path
 from typing import TYPE_CHECKING, Protocol
 
 # Backward-compat re-exports — moved to models.execution_request
-from vibe3.models.execution_request import (  # noqa: F401
+from vibe3.models import (  # noqa: F401
     ExecutionLaunchResult,
     ExecutionRequest,
+    WorktreeRequirement,  # noqa: F401
 )
-from vibe3.models.worktree import WorktreeRequirement  # noqa: F401
 
 if TYPE_CHECKING:
     from vibe3.agents.backends.async_launcher import AsyncExecutionHandle
