+++ src/vibe3/execution/role_interfaces.py
@@ -10,7 +10,7 @@ from __future__ import annotations
 from typing import Any, Callable, Protocol, runtime_checkable
 
 # Import SessionRole from canonical source instead of redefining
-from vibe3.models.session_types import SessionRole
+from vibe3.models import SessionRole
 
 
 @runtime_checkable
