+++ src/vibe3/execution/role_request_factory.py
@@ -9,10 +9,8 @@ from vibe3.execution.issue_role_support import (
     build_issue_async_cli_request,
     build_issue_sync_prompt_request,
 )
-from vibe3.models import IssueInfo
-from vibe3.models.execution_request import ExecutionRequest
-from vibe3.models.orchestra_config import OrchestraConfig
-from vibe3.services.convention_resolver import ConventionResolver
+from vibe3.models import ExecutionRequest, IssueInfo, OrchestraConfig
+from vibe3.services import ConventionResolver
 
 
 def build_role_async_request(
