+++ src/vibe3/execution/session_service.py
@@ -4,11 +4,10 @@ import re
 
 from loguru import logger
 
-from vibe3.clients.git_client import GitClient
-from vibe3.clients.sqlite_client import SQLiteClient
+from vibe3.clients import GitClient, SQLiteClient
 from vibe3.environment.session_registry import SessionRegistryService
 from vibe3.exceptions import SystemError, UserError
-from vibe3.models.session_types import SessionRole
+from vibe3.models import SessionRole
 
 _SESSION_ID_RE = re.compile(r"^[A-Za-z0-9_-]+$")
